reduction1 is fixed (hopefully for the final time)
We were extracting the same data segment each time (forgetting to increment by i). I assure you that it was not like this before it was in function form though.

Anyway, with the addition of the backProp function, learnMotion3 has been created, which applies the function to do its learning.