backProp3Layer was not really modified much, just comments and comment removal.
errorFunction was modified to accommodate for vector arguments.

classify3Layer and classify3LayerStrict apply the 3 layer neural network on an input. Strict simply forces the outputted value to be 0 or 1.

learnMotion4 uses a new reduction, reduction3, as well as Doris's OVO reader in order to learn on all of Alex's data at once (instead of just a few files at a time).

learnMotion4 creates a confusion matrix called confusionM. The goal here was to find a good reduction. We proved using learnMotion4 that reduction.m, the original one, was not going to work. Instead, reduction3.m computes the mean frequency {via weighted sum}, std dev of the frequency {via weighted sum}, and the energy of the signal. This results in a better confusion matrix, but may not be reliable, as a reduction either.