Use checkReduction to make a 3D plot of your reduction's separation. (assumes the reduction reduces to 3 params, of course).
Then run prepareData to produce randomized test and training data.

The files are put into ReducedData. One would load data_Alex_red3_learnable.mat in order to learn. The variables are found in the cells of TRAIN and TEST. The original reduced data is found in C.

TRAIN{1} = matrix for training walking data